(#128) refactor: 오타 수정 및 목데이터 수정
